ServerComms: allow user to suppress warnings about server errors.
So if the server is unavailable or if authentication fails, then they won't
have to see an error dialog each time they launch the program. The suppression
resets when authentication tokens are reset.

The server comms config dialog now always allows the user to clear their
settings, instead of only allowing it when successfully authenticated with the
server.