Notifications: use a more-unique thread ID.
It was potentially possible for the same thread ID to show up multiple times
if the user exited and then restarted the program, which could cause
confusing results. We now keep track of a more unique ID based on the thread
ID and the current system time, instead of using Thread.getId().